muracaat

+++ src/java/dao/MuracaatDAO.java
@@ -99,7 +99,7 @@ public class MuracaatDAO extends DBConnection {
                         rs.getDate("KAYIT_TARIHI"),
                         rs.getInt("AKTIF"),
                         rs.getDate("GUNCELLEME_TARIHI"),
-                        rs.getBigDecimal("KIMLIK_NO").toBigInteger(),
+                        rs.getInt("KIMLIK_NO"),
                         rs.getString("ISIM"),
                         rs.getString("SOYISIM"),
                         rs.getInt("ARSIV_DOSYA_NO"),
